requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
markdown==3.4.4
django-filter==23.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pytest-django==4.11.1
requests==2.32.3
django-extensions==3.2.1